Change balance amount type to String; improve tests
