+++ crates/starknet-devnet-server/src/api/http/endpoints/accounts.rs
@@ -6,7 +6,7 @@ use starknet_types::rpc::transaction_receipt::FeeUnit;
 
 use super::mint_token::{get_balance, get_erc20_address};
 use crate::api::http::error::HttpApiError;
-use crate::api::http::models::{Balance, SerializableAccount};
+use crate::api::http::models::{AccountBalanceResponse, SerializableAccount};
 use crate::api::http::{HttpApiHandler, HttpApiResult};
 
 pub async fn get_predeployed_accounts(
@@ -39,7 +39,7 @@ pub struct BalanceQuery {
 pub async fn get_account_balance(
     Extension(state): Extension<HttpApiHandler>,
     Query(params): Query<BalanceQuery>,
-) -> HttpApiResult<Json<Balance>> {
+) -> HttpApiResult<Json<AccountBalanceResponse>> {
     let account_address = ContractAddress::new(params.address)
         .map_err(|e| HttpApiError::InvalidValueError { msg: e.to_string() })?;
     let unit = params.unit.unwrap_or(FeeUnit::WEI);
@@ -49,5 +49,7 @@ pub async fn get_account_balance(
 
     let amount = get_balance(&mut starknet, account_address, erc20_address)
         .map_err(|e| HttpApiError::GeneralError(e.to_string()))?;
-    Ok(Json(Balance { amount, unit }))
+    let amount_str = amount.to_string();
+    println!("DEBUG amount_str: {amount_str}");
+    Ok(Json(AccountBalanceResponse { amount: amount_str, unit }))
 }
